Improving DetailProductScreen 3 and Adding Textchip

+++ app/src/main/java/com/kaizen/bangunpc/utils/Helper.kt
@@ -6,5 +6,5 @@ import java.util.Locale
 fun Int.toRupiahFormat(): String {
     val localeID = Locale("in", "ID")
     val formatRupiah = NumberFormat.getCurrencyInstance(localeID)
-    return formatRupiah.format(this.toDouble())
+    return formatRupiah.format(this.toDouble().toInt())
 }
